+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Bugfix:
 
+- Updated `Request-WsaLocalUserPasswordComplexity`, and `Request-WsaLocalUserAccountLockout` to pass a version with the `-drift` option is used. [GH-135](https://github.com/vmware/powershell-module-for-vmware-cloud-foundation-password-management/pull/135)
 - Updated `Request-WsaPasswordExpiration`, `Request-WsaPasswordComplexity`, and `Request-WsaAccountLockout` to pass a version with the `-drift` option is used. [GH-131](https://github.com/vmware/powershell-module-for-vmware-cloud-foundation-password-management/pull/131)
 
 Enhancement: